Bug fix and ps2+xboxOne controller setup
PS2 (dualshock2) and Xbox one controller (NOT 360) are included. To change controller settings, game.js must be opened and change 'PS3' to 'PS2' or 'xboxOne'. Some code reading is encouraged. (P.S. PS2 Dpad isn't working very well, but it is not necessary for this game)

Also, fixed a bug that when player is moved toward where it is facing every beginning of the match, (player 0 going right or player 1 goint left) player sprite's animation is not updated. That is, players are stuck in to "still" motion while they are moving. If player stop or move left, then they start to work again. I fixed this bug by setting both players in "idle" position. Animation only occurs when they are idle or opposite direction.